requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
joblib==1.1.0
numpy==1.22.0
Pillow==9.0.0
PyJWT==2.3.0
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tzdata==2021.5
Unidecode==1.3.2
whitenoise==5.3.0
